consumer_grafos/requirements.txt
pika==1.3.0
mysql-connector-python
